% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mantelNSGA2.R
\name{plot.mantelNSGA2}
\alias{plot.mantelNSGA2}
\title{plot results from mantelNSGA2}
\usage{
\method{plot}{mantelNSGA2}(
  obj,
  xlim = c(-1, 0),
  ylim = c(0, 1),
  genCol = hcl.colors(12, "viridis", alpha = 0.3),
  genPch = ".",
  genCex = 3,
  parFrontCol = 1,
  parFrontBg = 2,
  parFrontPch = 21,
  parFrontCex = 1,
  parFrontT = "p"
)
}
\arguments{
\item{obj}{bla}

\item{xlim}{bla}

\item{ylim}{bla}

\item{genCol}{bla}

\item{genPch}{bla}

\item{genCex}{bla}

\item{parFrontCol}{bla}

\item{parFrontBg}{bla}

\item{parFrontPch}{bla}

\item{parFrontCex}{bla}

\item{parFrontT}{bla}
}
\value{
a plot
}
\description{
plot results from mantelNSGA2
}
\examples{

library(vegan)
data("varechem")
data("varespec")

### envbio
# biological community variables that best correlate with environmental data
fix.mat = varechem
var.mat = wisconsin(varespec)
fix.dist.method = "euclidean"
var.dist.method = "bray"
scale.fix = TRUE
scale.var = FALSE
p.feat <- 0.1
2^ncol(var.mat)-1 # total combinations tested by bioEnv()

# mantelNSGA2
set.seed(1111)
fitGA <- mantelNSGA2(
  fix.mat = fix.mat, var.mat = var.mat,
  fix.dist.method = fix.dist.method, var.dist.method = var.dist.method,
  scale.fix = scale.fix, scale.var = scale.var, p.feat = p.feat,
  pop.size = 50, max.iter = 100, stop.criterion = 50,
  mutation.rate = 0.2, crossover.rate = 0.8
)
plot(fitGA$generation.fitness)
fitGA$pareto.solution
plot(fitGA, parFrontT = "o")

}
